project rearrangement
