+++ docs/map_split.md
@@ -1,7 +1,7 @@
 # 拆分与组合（用于 Git）
 
-> 模块名：`yrmap.gitutil`  
-> API 已在`yrmap`包中直接提供。无需再导入模块。
+> 模块名：`map_split`  
+> API 已在`pyalert2yr`包中直接提供。无需再导入模块。
 
 地图作为文本文件，很适合用 Git 记录差异。相比起“复制、重命名”的日期备份，Git 可以更快更好地帮助你“回溯”，甚至定位问题所在。
