--- yrmap/__init__.py
@@ -1,9 +1,0 @@
-# -*- encoding: utf-8 -*-
-# @File   : __init__.py
-# @Time   : 2023/12/30 19:29:01
-# @Author : Chloride
-from .gitutil import splitMap, joinMap
-
-__all__ = (
-    'splitMap', 'joinMap'
-)
